Rename extract amstercopy file
