+++ packages/contracts/hardhat.config.ts
@@ -12,7 +12,7 @@ const { INFURA_KEY, MNEMONIC, ETHERSCAN_API_KEY, PK } = process.env;
 
 import "./src/tasks/setup";
 import "./tasks/mastercopy-deploy";
-import "./tasks/mastercopy-store";
+import "./tasks/mastercopy-extract";
 import "./tasks/mastercopy-verify";
 
 const DEFAULT_MNEMONIC =
